backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
supabase==2.13.0
razorpay==1.4.2
PyJWT==2.10.1
gunicorn==23.0.0
